BACKEND/requirements.txt
Django==5.2.3
djangorestframework==3.15.1
mysqlclient==2.2.7
